+++ src/components/goal/HabitStats.vue
@@ -9,7 +9,7 @@ import { formatDate } from 'src/lib/date.ts';
 
 import StatTile from './StatTile.vue';
 import { commaify } from 'src/lib/number.ts';
-import { GoalHabitParameters } from 'server/lib/models/goal';
+import { GoalHabitParameters } from 'server/lib/models/goal.ts';
 import { GOAL_CADENCE_UNIT_INFO } from 'src/lib/goal.ts';
 
 const props = defineProps<{