+++ CHANGELOG.md
@@ -13,7 +13,9 @@ Types of changes include:
 ## Upcoming/Unreleased
 
 - NEW: Goal stats on the dashboard now show when you've completed your habit for the current period, streaks, and if you've hit your target.
+- NEW: Added a Total column to the table on the project page.
 - CHANGED: Habit goal pages are now nicer to look at and feature some stats.
+- FIXED: The chart on the project page will now always stay on the same tab when progress is added.
 
 ## 0.9.7