+++ CHANGELOG.md
@@ -12,6 +12,8 @@ Types of changes include:
 
 ## Upcoming/Unreleased
 
+## 0.9.8
+
 - NEW: You can now set your project total when entering and editing progress, as opposed to just entering the increment since last time.
 - NEW: Goal stats on the dashboard now show when you've completed your habit for the current period, streaks, and if you've hit your target.
 - NEW: Added a Total column to the table on the project page.